+++ lib/clean_framework.dart
@@ -10,5 +10,7 @@ export 'package:clean_framework/src/feature_state/feature_widget.dart';
 export 'package:clean_framework/src/logger.dart';
 export 'package:clean_framework/src/open_feature/open_feature.dart';
 export 'package:clean_framework/src/routing/app_router.dart';
+export 'package:clean_framework/src/utilities/deserializer.dart';
+export 'package:clean_framework/src/utilities/file.dart';
 export 'package:clean_framework/src/widgets/widgets.dart';
 export 'package:either_dart/either.dart';