+++ lib/src/defaults/providers/rest/rest_logger.dart
@@ -45,7 +45,10 @@ class _RequestLogger extends NetworkLogger {
 
   @override
   void initialize() {
-    printHeader('REQUEST', request.url.replace(queryParameters: {}).toString());
+    printHeader(
+      'REQUEST (${request.method})',
+      request.url.replace(queryParameters: {}).toString(),
+    );
     _printParams();
     _printPayload();
     _printHeaders();